+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - PHP 8.2 support
 - JSON formatted update requests
+- Solarium\Component\Highlighting\Highlighting::setQueryFieldPattern()
 
 
 ## [6.2.7]